# controller/v1/gin_request.go
package v1

import (
	"github.com/google/uuid"
	"pdf_service_api/models"
)

type GetDocumentRequest struct {
	DocumentUUID *uuid.UUID `json:"documentUUID"`
	OwnerUUID    *uuid.UUID `json:"ownerUUID"`
}

type CreateRequest struct {
	Base64        string     `json:"base64"`
	DocumentTitle *string    `json:"documentTitle"`
	OwnerUUID     *uuid.UUID `json:"ownerUUID"`
	OwnerType     *int       `json:"ownerType"`
}

type AddNewSelectionRequest struct {
	DocumentUUID    *uuid.UUID                        `json:"documentUUID,omitempty"`
	IsComplete      bool                              `json:"isComplete,omitempty"`
	Settings        *string                           `json:"settings,omitempty"`
	SelectionBounds *map[int][]models.SelectionBounds `json:"selectionBounds,omitempty"`
}

type AddMetaRequest struct {
	NumberOfPages *uint32
	Height        *float32
	Width         *float32
	Images        *map[uint32]string
}

type UpdateMetaRequest struct {
	UUID          uuid.UUID
	NumberOfPages *uint32
	Height        *float32
	Width         *float32
	Images        *map[uint32]string
}

type DeleteMetaRequest struct {
	UUID uuid.UUID
}

type GenerateMetaRequest struct {
	Base64 string `json:"base64"`
}