# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The API documentation for the golang backend server.",
        "title": "Go Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/documents": {
            "get": {
                "description": "Retrieves document details. Documents can be fetched either by their unique Document UUID or by an Owner UUID.\nOptional exclusion parameters can be used to omit specific fields from the response.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "documents"
                ],
                "summary": "Get documents",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The unique identifier of the document to retrieve. If provided",
                        "name": "documentUUID",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "The unique identifier of the owner whose documents are to be retrieved.",
                        "name": "ownerUUID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "Fields to exclude from the response. Allowed values: `documentTitle`, `timeCreated`, `ownerUUID`, `ownerType`, `base64`.",
                        "name": "exclude",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved document(s).",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "documents": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/models.Document"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request: Invalid UUID format or no valid parameters specified.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found: No document(s) found for the given UUID.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error: An unexpected error occurred on the server.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Uploads a document by receiving its base64 encoded string in the request body.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "documents"
                ],
                "summary": "Upload a new document",
                "parameters": [
                    {
                        "description": "Document upload request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful upload, returns the document UUID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request, typically due to invalid input or upload failure"
                    }
                }
            },
            "delete": {
                "description": "Deletes a document based on the provided document UUID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "documents"
                ],
                "summary": "Delete a document",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The UUID of the document to delete",
                        "name": "documentUUID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The UUID of the owner of the document that is getting deleted",
                        "name": "ownerUUID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful deletion",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request, typically due to missing/invalid UUID or deletion failure"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Document": {
            "type": "object",
            "properties": {
                "base64": {
                    "type": "string"
                },
                "documentTitle": {
                    "type": "string"
                },
                "documentUUID": {
                    "type": "string",
                    "example": "ba3ca973-5052-4030-a528-39b49736d8ad"
                },
                "ownerType": {
                    "type": "integer"
                },
                "ownerUUID": {
                    "type": "string"
                },
                "selectionData": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Selection"
                    }
                },
                "timeCreated": {
                    "type": "string"
                }
            }
        },
...
        "models.Selection": {
            "type": "object",
            "properties": {
                "documentUUID": {
                    "type": "string"
                },
                "isComplete": {
                    "type": "boolean"
                },
                "selectionBounds": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/models.SelectionBounds"
                        }
                    }
                },
                "selectionUUID": {
                    "type": "string"
                },
                "settings": {
                    "type": "string"
                }
            }
        },
        "models.SelectionBounds": {
            "type": "object",
            "properties": {
                "extract_method": {
                    "type": "string",
                    "example": "None"
                },
                "x1": {
                    "type": "number",
                    "example": 43.122
                },
                "x2": {
                    "type": "number",
                    "example": 13
                },
                "y1": {
                    "type": "number",
                    "example": 52.125
                },
                "y2": {
                    "type": "number",
                    "example": 27.853
                }
            }
        },
...
        "v1.CreateRequest": {
            "type": "object",
            "properties": {
                "base64": {
                    "type": "string"
                },
                "documentTitle": {
                    "type": "string"
                },
                "ownerType": {
                    "type": "integer"
                },
                "ownerUUID": {
                    "type": "string"
                }
            }
        },
...
    },
...
}