# models/document.go
package models

import (
	"github.com/google/uuid"
	"time"
)

type Document struct {
	Uuid          uuid.UUID    `json:"documentUUID" example:"ba3ca973-5052-4030-a528-39b49736d8ad"`
	DocumentTitle *string      `json:"documentTitle,omitempty"`
	TimeCreated   *time.Time   `json:"timeCreated,omitempty"`
	OwnerUUID     *uuid.UUID   `json:"ownerUUID,omitempty"`
	OwnerType     *int         `json:"ownerType,omitempty"`
	Base64        *string      `json:"base64,omitempty"`
	SelectionData *[]Selection `json:"selectionData,omitempty"`
}

type DocumentRepository interface {
	UploadDocument(document Document) error
	GetDocumentByDocumentUUID(document, owner uuid.UUID, excludes map[string]bool) (Document, error)
	GetDocumentByOwnerUUID(owner uuid.UUID, limit int8, offset int8, excludes map[string]bool) ([]Document, error)
	DeleteDocumentById(documentUuid, ownerUuid uuid.UUID) error
}